Draw Filled Triangle and change draw methods to take ints
Previously we were taking a Triangle or Vector into the draw
functions. Those types use float64s and it was causing the math to be
subtley off and there were artifacts in the DrawFilledTriangle
surfaces.

We can probably change this back in the future and handle it better,
but for now lets keep moving.